refactor: extract followup analysis hooks/helpers
