+++ src/pages/FollowUpAnalysis/convert-element-to-un-followup-request.js
@@ -1,0 +1,10 @@
+const convertElementToUnFollowupRequest = e => ({
+    dataElementId: e.dataElementId,
+    periodId: e.periodId,
+    organisationUnitId: e.organisationUnitId,
+    categoryOptionComboId: e.categoryOptionComboId,
+    attributeOptionComboId: e.attributeOptionComboId,
+    followup: false,
+})
+
+export default convertElementToUnFollowupRequest
